bug fix and small additions
Fixed a crucial bug where the config could have been overwritten with empty file (that's how we say to "accidentally deleted")

I am testing the idea of using a fan instead of a heater. so, why not both. Fans are 10x more energy efficient and seem to solve the problem with heat spikes more reliable than the heater.

further reorganization of the shared functions.

updated install routine and log.